% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcpot.R
\name{calcpot}
\alias{calcpot}
\title{Cálculo de las potencias de una matriz 
cuadrada mediante la propiedad de semejanza
de matrices P^-1*A*P = D <=> A = P*D*P^-1}
\usage{
calcpot(A, n)
}
\description{
Cálculo de las potencias de una matriz 
cuadrada mediante la propiedad de semejanza
de matrices P^-1*A*P = D <=> A = P*D*P^-1
}
