% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detespectral.R
\name{detespectral}
\alias{detespectral}
\title{Función que calcula el determinante de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices}
\usage{
detespectral(A)
}
\description{
Función que calcula el determinante de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices
}
