% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gramschmidt.R
\name{gramschmidt}
\alias{gramschmidt}
\title{Funcion que calcula el Algoritmo QR de una matriz
cuadrada A mostrando por pantalla las matrices Q y R}
\usage{
gramschmidt(A)
}
\description{
Funcion que calcula el Algoritmo QR de una matriz
cuadrada A mostrando por pantalla las matrices Q y R
}
