% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrizdiag.R
\name{matrizdiag}
\alias{matrizdiag}
\title{Función que calcula la matriz diagonal D
a partir de los valores propios de A}
\usage{
matrizdiag(A)
}
\description{
Función que calcula la matriz diagonal D
a partir de los valores propios de A
}
