% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trazaespectral.R
\name{trazaespectral}
\alias{trazaespectral}
\title{Función que calcula la traza de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices}
\usage{
trazaespectral(A)
}
\description{
Función que calcula la traza de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices
}
