% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valprop.R
\name{valprop}
\alias{valprop}
\title{Función que calcula los valores propios de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices}
\usage{
valprop(A)
}
\description{
Función que calcula los valores propios de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices
}
