% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecprop.R
\name{vecprop}
\alias{vecprop}
\title{Función que calcula los vectores propios de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices}
\usage{
vecprop(A)
}
\description{
Función que calcula los vectores propios de una matriz cuadrada
y muestra el resultado por pantalla utilizando la Teoría
Espectral de Matrices
}
